Move `tests` to `video_compositor` crate
